+++ variant_topic_tools/include/variant_topic_tools/CollectionVariant.tpp
@@ -22,11 +22,11 @@ namespace variant_topic_tools {
 /* Accessors                                                                 */
 /*****************************************************************************/
 
-template <typename T> T& CollectionVariant::getValue(size_t index) {
+template <typename T> T& CollectionVariant::getValue(int index) {
   return this->getMember(index).template getValue<T>();
 }
 
-template <typename T> const T& CollectionVariant::getValue(size_t index)
+template <typename T> const T& CollectionVariant::getValue(int index)
     const {
   return this->getMember(index).template getValue<T>();
 }
