+++ variant_topic_tools/include/variant_topic_tools/MessageFieldCollection.h
@@ -37,41 +37,41 @@ namespace variant_topic_tools {
   template <typename T> class MessageFieldCollection {
   public:
     /** \brief Default constructor
-      */ 
+      */
     MessageFieldCollection();
-    
+
     /** \brief Copy constructor
-      */ 
+      */
     MessageFieldCollection(const MessageFieldCollection<T>& src);
-    
+
     /** \brief Destructor
-      */ 
+      */
     virtual ~MessageFieldCollection();
-    
+
     /** \brief Retrieve the number of fields of the message field collection
       */
     size_t getNumFields() const;
-    
+
     /** \brief Retrieve a field of the message field collection by index
       *   (non-const version)
       */
-    MessageField<T>& getField(size_t index);
-    
+    MessageField<T>& getField(int index);
+
     /** \brief Retrieve a field of the message field collection by index
       *   (const version)
       */
-    const MessageField<T>& getField(size_t index) const;
-    
+    const MessageField<T>& getField(int index) const;
+
     /** \brief Retrieve a field of the message field collection by name
       *   (non-const version)
       */
     MessageField<T>& getField(const std::string& name);
-    
+
     /** \brief Retrieve a field of the message field collection by name
       *   (const version)
       */
     const MessageField<T>& getField(const std::string& name) const;
-    
+
     /** \brief True, if the message field collection contains the
       *   field with the specified name
       */
@@ -84,88 +84,88 @@ namespace variant_topic_tools {
     /** \brief Append a field to the message field collection
       */
     void appendField(const MessageField<T>& field);
-    
+
     /** \brief Append a field to the message field collection (overloaded
       *   version taking a field name and a field value)
       */
     void appendField(const std::string& name, const T& value = T());
-    
+
     /** \brief Merge this message field collection with another message
       *   field collection
       */
     void merge(const MessageFieldCollection<T>& collection);
-    
+
     /** \brief Clear the message field collection
       */
     virtual void clear();
-    
+
     /** \brief Write the message field collection to a stream
       */
     virtual void write(std::ostream& stream, const std::string& indent =
       std::string()) const;
-      
+
     /** \brief Operator for retrieving the fields of the message field
       *   collection by index (non-const version)
       */
-    MessageField<T>& operator[](size_t index);
-    
+    MessageField<T>& operator[](int index);
+
     /** \brief Operator for retrieving the fields of the message field
       *   collection by index (const version)
       */
-    const MessageField<T>& operator[](size_t index) const;
-    
+    const MessageField<T>& operator[](int index) const;
+
     /** \brief Operator for retrieving the fields of the message field
       *   collection by name (non-const version)
       */
     MessageField<T>& operator[](const std::string& name);
-    
+
     /** \brief Operator for retrieving the fields of the message field
       *   collection by name (const version)
       */
     const MessageField<T>& operator[](const std::string& name) const;
-    
+
     /** \brief Operator for appending a field to the message field collection
       */
     MessageFieldCollection<T>& operator+=(const MessageField<T>& field);
-    
+
     /** \brief True, if this message field collection is equal to another
       *   message field collection
       */
     bool operator==(const MessageFieldCollection<T>& collection) const;
-    
+
     /** \brief True, if this message field collection is not equal to another
       *   message field collection
       */
     bool operator!=(const MessageFieldCollection<T>& collection) const;
-    
+
   protected:
     /** \brief Declaration of the message field pointer type
       */
     typedef boost::shared_ptr<MessageField<T> > MessageFieldPtr;
-    
+
     /** \brief Declaration of the message field weak pointer type
       */
     typedef boost::weak_ptr<MessageField<T> > MessageFieldWPtr;
-    
+
     /** \brief The message fields of the collection in order
       */
     std::vector<MessageFieldPtr> fieldsInOrder;
-    
+
     /** \brief The message fields of the collection by name
       */
     boost::unordered_map<std::string, MessageFieldPtr> fieldsByName;
-    
+
     /** \brief Recursively retrieve a field of the message field collection
       *   by name
       */
     MessageField<T>& getField(const std::string& name, size_t pos) const;
-    
+
     /** \brief True, if the message field collection or any of its fields
       *   contains the field with the specified name
       */
     bool hasField(const std::string& name, size_t pos) const;
   };
-  
+
   /** \brief Operator for writing the message field collection to a stream
     */
   template <typename T> std::ostream& operator<<(std::ostream& stream,
